% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{abstract}
\alias{abstract}
\title{abstract}
\usage{
abstract(..., cache_name = NULL, sep = "\\n\\n")
}
\arguments{
\item{...}{headed abstract sections}

\item{cache_name}{character indicating the name the cached r chunk}

\item{sep}{how should sections be pasted together?}
}
\description{
Create a structured abstract
}
