% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_comments.R
\name{bullets_2_numbers}
\alias{bullets_2_numbers}
\title{bullets_2_numbers}
\usage{
bullets_2_numbers(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
Convert bullets to numbered lists
}
\examples{
bullets_2_numbers(c("* First bullet\n\n* Second bullet\n\n* Third bullet\n\n"))
}
