% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_manuscript.R
\name{extract_html_sections}
\alias{extract_html_sections}
\title{extract_html_sections}
\usage{
extract_html_sections(path)
}
\arguments{
\item{path}{character path to file}
}
\description{
A function to extract text within spans
}
