% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_manuscript.R
\name{extract_md_sections}
\alias{extract_md_sections}
\title{extract_sections}
\usage{
extract_md_sections(path)

extract_md_sections(path)
}
\arguments{
\item{path}{character path to file}
}
\description{
Extracts markdown sections

Extracts markdown sections
}
