% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nested_md.R
\name{extract_md_sections2}
\alias{extract_md_sections2}
\title{extract_md_sections2}
\usage{
extract_md_sections2(string)
}
\arguments{
\item{string}{string to extract md sections from}
}
\description{
Extract text when there are nested parentheses
}
