% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nested_md.R
\name{extract_sections}
\alias{extract_sections}
\title{extract_sections}
\usage{
extract_sections(string, is_span = FALSE)
}
\arguments{
\item{string}{string to extract md sections from}

\item{is_span}{logical, whether to extract sections in square brackets (default) or spans}
}
\description{
Extract sections from text string
}
