% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_pages}
\alias{find_pages}
\title{find_pages}
\usage{
find_pages(manuscript, string)
}
\arguments{
\item{manuscript}{the manuscript object to search}

\item{string}{the string to find}
}
\description{
Searches through a manuscript for a page reference
}
