% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_output.r
\name{letter.docx}
\alias{letter.docx}
\title{letter.docx}
\usage{
letter.docx(...)
}
\arguments{
\item{...}{additional arguments passed to rmarkdown::output_format}
}
\description{
Revise and resubmit letter as a txt doc
}
