% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_output.R
\name{letter.txt}
\alias{letter.txt}
\title{letter.txt}
\usage{
letter.txt(...)
}
\arguments{
\item{...}{additional arguments passed to rmarkdown::output_format}
}
\description{
Revise and resubmit letter as a txt doc
}
