% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\title{FUNCTION_TITLE}
\usage{
mask(
  x,
  mask_message = "masked for blind review",
  hide = rmarkdown::metadata$mask
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{mask_message}{PARAM_DESCRIPTION, Default: 'masked for blind review'}

\item{hide}{PARAM_DESCRIPTION, Default: rmarkdown::metadata$mask}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
mask("this is masked", hide = TRUE)
}
\seealso{
\code{\link[rmarkdown]{metadata}}
}
