% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_manuscript.R
\name{print.manuscript}
\alias{print.manuscript}
\title{print.manuscript}
\usage{
\method{print}{manuscript}(x, ...)
}
\arguments{
\item{x}{an object of class "manuscript"}

\item{...}{other arguments}
}
\description{
Print method for manuscripts
}
