% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_manuscript.R
\name{read_spans}
\alias{read_spans}
\title{read_spans}
\usage{
read_spans(address, id)
}
\arguments{
\item{address}{string url or file location of knited html}

\item{id}{the id from a html tag}
}
\description{
read_spans
}
