% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manuscripts.R
\name{remove_manuscript}
\alias{remove_manuscript}
\title{Remove manuscript from environment}
\usage{
remove_manuscript(manuscript = NULL, envir = parent.frame(1))
}
\arguments{
\item{manuscript}{A character vector of manuscript names
that exist in \code{envir}.}

\item{envir}{The environment from which to remove
the manuscript.}
}
\value{
Invisibly returns \code{TRUE} if successful, and \code{FALSE}
otherwise.
}
\description{
Removes a manuscript from a given environment by name.
}
\examples{
remove_manuscript()
}
