% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revise.R
\docType{package}
\name{revise}
\alias{revise}
\title{revise: Tools for Formulating Revise and Resubmit Rmarkdown Documents}
\description{
When writing scientific manuscripts, it is challenging to keep revisions
synchronized between the manuscript and the action letter detailing
responses to reviewers.
}
\details{
Revise automates this process. Sections in the manuscript can be tagged,
and inserted into the action letter.
}
