% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_output.r
\name{revise_letter_docx}
\alias{revise_letter_docx}
\title{Dynamic docx Revision Letter}
\usage{
revise_letter_docx(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[bookdown:html_document2]{bookdown::word_document2()}}.}
}
\value{
An R Markdown output format object to be passed to
  \code{rmarkdown::\link[rmarkdown]{render}()}.
}
\description{
Template for creating journal revision letters
with dynamic excerpts from a manuscript.
}
\details{
This function wraps \code{\link[bookdown:html_document2]{bookdown::word_document2()}}.
}
\seealso{
\code{\link[bookdown:html_document2]{bookdown::word_document2()}}
}
