% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_output.r
\name{revise_letter_txt}
\alias{revise_letter_txt}
\title{Dynamic txt Revision Letter}
\usage{
revise_letter_txt(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[rmarkdown:output_format]{rmarkdown::output_format()}}.}
}
\value{
An R Markdown output format definition that can be passed to
  \code{\link[rmarkdown]{render}}.
}
\description{
Template for creating journal revision letters
with dynamic excerpts from a manuscript.
}
\details{
This function wraps \code{\link[rmarkdown:output_format]{rmarkdown::output_format()}}.
}
\seealso{
\code{\link[rmarkdown:output_format]{rmarkdown::output_format()}}
}
