% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{drive_comments}
\alias{drive_comments}
\title{drive_comments}
\usage{
drive_comments(url, email = NULL, cache_credentials = NULL)
}
\arguments{
\item{url}{URL to drive document}

\item{email}{drive email account. Can be set with options(gargle_oath_email)}

\item{cache_credentials}{bool. Should cached credentials be used? Can be set with options(gargle_oauth_cache)}
}
\description{
Extract comments from a google docs document
}
