% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_inline}
\alias{evaluate_inline}
\title{evaluate_inline}
\usage{
evaluate_inline(string, envir = parent.frame(), trust_manuscript = FALSE)
}
\arguments{
\item{string}{a section of text containing inline \code{r} elements}

\item{envir}{the environment in which to evaluate the code}

\item{trust_manuscript}{logical. Should inline \code{r} code be trusted and
evaluated? Defaults to \code{FALSE}. When called via \code{\link[=get_revision]{get_revision()}}, the default
comes from \code{getOption("revise_trust_manuscript")}.}
}
\value{
A character string with inline \code{r} code evaluated.
}
\description{
Evaluates embedded inline r code within a string of text
}
\details{
Inline \code{r} code is detected using backticks and evaluated via
\code{\link[glue:glue]{glue::glue()}} with \verb{.open = "}r "\code{and}.close = "\verb{"}.
If inline \code{r} chunks are present and \code{trust_manuscript} is not \code{TRUE},
this function errors with the inline code that would have been evaluated.
Set \code{options(revise_trust_manuscript = TRUE)} and call \code{\link[=get_revision]{get_revision()}}, or pass
\code{trust_manuscript = TRUE} directly, or disable evaluation in
\code{\link[=get_revision]{get_revision()}} by setting \code{evaluate = FALSE}.
}
\examples{
revise:::evaluate_inline("1+1 = `r 1+1`")
}
