% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_manuscript.R
\name{evaluate_inline}
\alias{evaluate_inline}
\title{evaluate_inline}
\usage{
evaluate_inline(string)
}
\arguments{
\item{string}{a section of text with includes inline elements}
}
\description{
Evaluates embedded rchunks within a string of text
}
\examples{
revise:::evaluate_inline("1+1 = `r 1+1`")
}
