% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_parser.R
\name{extract_fenced_sections}
\alias{extract_fenced_sections}
\title{extract_fenced_sections}
\usage{
extract_fenced_sections(string, ignore_ids = c("refs"))
}
\arguments{
\item{string}{string to extract md sections from}

\item{ignore_ids}{character vector of ids to skip (default: "refs")}
}
\description{
Extract sections from fenced divs using markdown tags.
}
