% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_revision.R
\name{get_revision}
\alias{get_revision}
\title{get_revision}
\usage{
get_revision(
  id,
  manuscript = NULL,
  quote = TRUE,
  evaluate = TRUE,
  split_string = FALSE,
  search_length = 300,
  include_pgnum = TRUE
)
}
\arguments{
\item{id}{the id from a html tag}

\item{manuscript}{Optional; a manuscript object in which to search
for \code{id}. Defaults to \code{NULL}, in which case the manuscript object
is found in the environment.}

\item{quote}{is the output chunk quoted?}

\item{evaluate}{logical. Should inline rchunks be executed?}

\item{split_string}{should only the start and end of the string be searched for?}

\item{search_length}{numeric. Searches for the first n and n characters in a string. Shorten if difficult to find passages split by floats.}

\item{include_pgnum}{logical. include PDF page number?}
}
\description{
Extract and format revision
}
