% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_manuscript.R
\name{read_manuscript}
\alias{read_manuscript}
\title{read_manuscript}
\usage{
read_manuscript(address, id = NULL, PDF = NULL)
}
\arguments{
\item{address}{path to a rmarkdown file}

\item{id}{if provided spans will be searched for text which will be returned}

\item{PDF}{if TRUE, or path provided, a PDF will be loaded for page matching.}
}
\description{
Reads in rmarkdown manuscript and an associated PDF as possible
}
