% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revise_letter.R
\name{revise_letter_pdf}
\alias{revise_letter_pdf}
\title{Dynamic PDF Revision Letter}
\usage{
revise_letter_pdf(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[papaja:revision_letter_pdf]{papaja::revision_letter_pdf()}}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Template for creating journal revision letters
with dynamic excerpts from a manuscript.
}
\details{
This function wraps \code{\link[papaja:revision_letter_pdf]{papaja::revision_letter_pdf()}}.
}
\seealso{
\code{\link[papaja:revision_letter_pdf]{papaja::revision_letter_pdf()}}, \code{\link[bookdown:html_document2]{bookdown::pdf_document2()}}, \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}
}
